SAdd dynamic scrollbar level of abstraction.
