+++ src/plugins/timeline/components/dynamic-scrollbar.scss
@@ -1,11 +1,11 @@
-.timeline-scrollbar {
+.dynamic-scrollbar {
   width: 100%;
   height: 17px;
   background-color: #d9d9d9;
   position: relative;
   border-radius: 2px;
 
-  .timeline-scrollbar-thumb {
+  .dynamic-scrollbar-thumb {
     position: absolute;
     top: 0;
     height: 100%;
