+++ src/plugins/timeline/components/timeline-tile.tsx
@@ -1,10 +1,8 @@
 import { observer } from "mobx-react";
-import React, { useContext } from "react";
+import React from "react";
 import { BasicEditableTileTitle } from "../../../components/tiles/basic-editable-tile-title";
 import { ITileProps } from "../../../components/tiles/tile-component";
 import { TileToolbar } from "../../../components/toolbar/tile-toolbar";
-import { TileModelContext } from "../../../components/tiles/tile-api";
-import { isTimelineContentModel } from "../models/timeline-content";
 import { Timeline } from "./timeline";
 import { TimelineKey } from "./timeline-key";
 import { TimelineScrollbar } from "./timeline-scrollbar";
@@ -12,13 +10,6 @@ import "../timeline-toolbar";
 import "./timeline-tile.scss";
 
 export const TimelineComponent: React.FC<ITileProps> = observer(function TimelineComponent({ readOnly, tileElt }) {
-  const rawContent = useContext(TileModelContext)?.content;
-  const model = isTimelineContentModel(rawContent) ? rawContent : undefined;
-  if (!model) return null;
-
-  const { dataStartTime, dataEndTime, viewStartTime, viewEndTime, setViewRange } = model;
-  const showScrollbar = dataStartTime != null && dataEndTime != null && viewStartTime != null && viewEndTime != null;
-
   return (
     <div className="tile-content timeline-tile">
       <BasicEditableTileTitle />
@@ -30,15 +21,7 @@ export const TimelineComponent: React.FC<ITileProps> = observer(function Timelin
           <div className="event-label">Event</div>
         </div>
         <Timeline />
-        {showScrollbar && (
-          <TimelineScrollbar
-            dataStartTime={dataStartTime}
-            dataEndTime={dataEndTime}
-            viewStartTime={viewStartTime}
-            viewEndTime={viewEndTime}
-            onViewChange={setViewRange}
-          />
-        )}
+        <TimelineScrollbar />
         <TimelineKey />
       </div>
     </div>
